bugfix: math_valid

+++ math3dfunc.cpp
@@ -960,7 +960,7 @@ math3d_minmax(struct math_context *M, math_t transform, math_t points) {
 
 	if (math_isnull(transform)){
 		minmax[0] = minmax[1] = VEC(M, p);
-		for (int ii=1; ii<numpoints; ++ii){
+		for (int ii=1; ii<(int)numpoints; ++ii){
 			const auto& pp = VEC(M, math_index(M, points, ii));
 			minmax[0] = glm::min(minmax[0], pp);
 			minmax[1] = glm::max(minmax[1], pp);
@@ -968,7 +968,7 @@ math3d_minmax(struct math_context *M, math_t transform, math_t points) {
 	} else {
 		const glm::mat4& m = MAT(M, transform);
 		minmax[0] = minmax[1] = m * VEC(M, p);
-		for (int ii=1; ii<numpoints; ++ii){
+		for (int ii=1; ii<(int)numpoints; ++ii){
 			const glm::vec4 tpp = m * VEC(M, math_index(M, points, ii));
 			minmax[0] = glm::min(minmax[0], tpp);
 			minmax[1] = glm::max(minmax[1], tpp);
